pyproject.toml
[tool.poetry]
name = "django-rdcapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

django = "^5.0.4"
djangorestframework = "^3.15.1"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"
django-import-export = "^3.3.8"

djoser = "^2.2.2"

gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"

django-stubs-ext = "^4.2.7"
drf-spectacular = "^0.27.2"
numpy = "^1.26.4"
jsonschema = "^4.21.1"
django-jsonschema-form = "^1.0.3"
ruff = "^0.4.2"

[tool.poetry.group.dev.dependencies]
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"
django-debug-toolbar = "^4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
